fix: split AddWish component
