+++ src/components/AddWish/AddWish.scss
@@ -17,7 +17,7 @@
   max-width: 780px;
   padding: 30px;
   display: flex;
-  justify-content: center;
+  justify-content: space-between;
   align-items: center;
   flex-direction: column;
   background-color: $page-background-color;
@@ -25,7 +25,6 @@
     0 9px 46px 8px rgba(0, 0, 0, 0.12);
   overflow-y: scroll;
   height: calc(100% - 30px);
-  margin: 15px;
 }
 
 .addWish_content_header {
@@ -35,7 +34,6 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  padding-top: 50px;
 }
 
 .wish-window-cover {
@@ -164,50 +162,4 @@ div .uploadImg:hover {
 .addFileGradients {
   display: flex;
   flex-wrap: wrap;
-}
-
-.addImageUrlContainer {
-  display: none;
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgb(189, 189, 189, 0.7);
-  z-index: 90;
-}
-
-.addImageUrl {
-  width: 70%;
-  margin: 20vh auto;
-  padding: 10px;
-  background: rgb(255, 255, 255);
-  z-index: 100;
-}
-
-.headerUrl {
-  display: flex;
-  justify-content: space-between;
-}
-
-div .closeUrl {
-  font-size: 2.5rem;
-  cursor: pointer;
-}
-
-.closeUrl:hover {
-  color: red;
-  transition: all 0.3s;
-}
-
-.addImageUrl form {
-  width: 100%;
-}
-.badUrl {
-  margin-left: 10%;
-  color: rgb(255, 38, 38);
-  display: none;
-}
-.addImageUrl .submit_btns {
-  margin: 0;
 }
