+++ src/components/AddWishUrlContainer/AddWishUrlContainer.scss
@@ -1,0 +1,63 @@
+.addImageUrlContainer {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgb(189, 189, 189, 0.7);
+  z-index: 90;
+}
+.hide {
+  display: none;
+}
+
+.addImageUrl {
+  width: 70%;
+  margin: 20vh auto;
+  padding: 10px;
+  background: rgb(255, 255, 255);
+  z-index: 100;
+}
+
+.headerUrl {
+  display: flex;
+  justify-content: space-between;
+}
+
+div .closeUrl {
+  font-size: 2.5rem;
+  cursor: pointer;
+}
+
+.closeUrl:hover {
+  color: red;
+  transition: all 0.3s;
+}
+
+.addImageUrl form {
+  width: 100%;
+}
+.badUrl {
+  margin-left: 10%;
+  color: rgb(255, 38, 38);
+}
+
+.submit_btns {
+  margin: 25px;
+  display: flex;
+  justify-content: space-between;
+}
+
+.submit_btns button {
+  width: 95px;
+  height: 30px;
+  cursor: pointer;
+}
+
+.submit_btns button:last-child {
+  background: rgb(231, 76, 31);
+  border: none;
+}
+.addImageUrl .submit_btns {
+  margin: 0;
+}
