+++ src/components/AddWishUrlContainer/index.tsx
@@ -1,0 +1,3 @@
+import AddWishUrlContainer from '@/components/AddWishUrlContainer/AddWishUrlContainer';
+
+export default AddWishUrlContainer;
